refactor: move `CHAT_PERSONAS` to `bot.domain.chat` and remove obsolete `chat_personas.py` file for improved organization
